Impl navigate to TopAppBarType screen
