+++ app/src/main/AndroidManifest.xml
@@ -32,10 +32,6 @@
             android:label="@string/hide_on_scroll" />
 
         <activity
-            android:name=".components.appbar.top.TopAppBarTypeActivity"
-            android:label="@string/top_app_bar" />
-
-        <activity
             android:name=".components.appbar.top.ActionBarActivity"
             android:label="@string/actionbar" />
